requirements.txt
# Resume Parsing Framework Dependencies

# PDF parsing
pypdf>=3.0.0

# Word document parsing
python-docx>=1.1.0

# Google Gemini LLM
google-genai>=0.2.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

